did sum refactoring for ease later when i make ai stuff
