+++ Commands/gRole.js
@@ -1,4 +1,4 @@
-const { setGrole } = require("../helperFunc.js");
+const { setGrole } = require('../HelperFunctions/adminHelpers.js');
 const { SlashCommandBuilder } = require('@discordjs/builders');
 
 module.exports = {
