+++ Commands/setVote.js
@@ -1,4 +1,4 @@
-const { setVote } = require("../helperFunc.js");
+const { setVote } = require('../HelperFunctions/adminHelpers.js');
 const { SlashCommandBuilder } = require('@discordjs/builders');
 
 module.exports = {
